MP fix for A-10-bombable
Filename for AI/Aircraft/A-10-Bombable/Models/A-10.xml didn't
match the filename for Aircraft/A-10-Bombable/Models/A-10-model.xml

So when FG was flown with A-10-Bombable over MP it picked up the Aircraft/ version
rather than the AI/Aircraft/ version.

The Aircraft/version lacks the Bombable code, so MP/AI didn't work.

Changing the filename from A-10.xml to A-10-model.xml fixes the problem.

Code in A-10-Bombable-set.xml was changed to match.