Sopwith Camel - Fix to sense/gyroscopic issue and various tweaks
JSBSim gyroscopic moment reversal is dealt with a new way, by reversing the Sense to get proper direction of the gyroscopic moment (but reversed P Factor and Torque) and then introducing a correction to the torque to reverse it.  This works better; P Factor is still incorrect but this is far less noticeable.

Experimented with a smaller varue for Gyroscopic moment from a historical calculation but I believe my calculation is more accurate.

Tweaked drag and critical angle of attack values to try to duplicate the famed 'speedy turn' of the Camel, supposedly up to 72 degree per second.  This model is currently lower than that.